feat: unify & streamline Alpadreams bidirectional stack (model, pipeline, workflow, and conversion)
- Consolidated all bidirectional Alpadreams functionality into a single entry point (`run_alpadreams.py`), deprecating legacy teacher scripts and code paths. Updated configs and code to use consistent `AlpadreamsPipeline` and `AlpadreamsPipelineConfig` naming throughout.
- Modernized the core workflow: replaced legacy Cosmos teacher and SIL HDMap implementations with a unified Cosmos transformer featuring classifier-free guidance and the UniPC scheduler, supporting fully bidirectional inference and training.
- Refactored video processing for clarity and efficiency: consolidated frame handling (removing previous FPS/frame logic and unused routines), clarified resizing, integrated robust video reading into the main workflow, and enforced clear input/output pixel dimension requirements for reproducibility.
- Enhanced configuration, argument parsing, and error handling, including overwrite config name logic, improved builder/validation for inference/conversion, GCP path overrides, and more informative checkpoint loading with stricter error checks (raises on unchanged distributed state-dict entries).
- Added robust support for negative prompts and embedding workflows:
    - Introduced a dedicated constants module for the negative prompt.
    - Pipeline now auto-infers negative text prompt requirements (when guidance scale > 1), and only saves/loads negative embeddings as needed for memory efficiency.
    - Refactored the negative prompt handling logic for consistency and maintainability across all code paths.
- Streamlined and standardized model conversion into a single `convert_i4_dcp2pt.py` script with a clean, consistent interface and credential handling; updated all documentation (README, in-code comments) to match new instructions, workflow, and file naming/output conventions.
- Improved logging: reduced redundancy in video chunk/frame and AR-step output, focusing on key info for easy debugging and maintainability.
- Comprehensive test updates:
    - Added/extended unit tests to validate new negative prompt and embedding handling (with mocks, for both present/absent negative embeddings).
    - Extended bidirectional tests to ensure negative prompt flow is correctly required and wired.
- Refined import structure, clarified inline documentation, improved maintainability, and recorded all major deprecations, transitions, and refactors for future reference.

+++ flashdreams/flashdreams/infra/diffusion/scheduler/fm_unipc.py
@@ -328,7 +328,12 @@ class FlowMatchUniPCScheduler(Scheduler):
         last_sample: Tensor | None = None
 
         for i in range(N):
-            timestep = self.timesteps[i]
+            # Schedule buffers are pinned to fp32 (to preserve integer
+            # timestep values under a stray `module.to(bf16)`), but the
+            # network expects timesteps in the input dtype so that
+            # downstream modulation / Linear layers stay consistent.
+            timestep = self.timesteps[i].to(dtype=input_dtype)
+
             # Network forward (heavy compute -- everything else here is
             # ~free relative to this).
             flow = predict_flow(sample, timestep)
